added ajax functionality and server api

+++ controllers/comments_controller.js
@@ -16,21 +16,39 @@ module.exports.create = async (req, res) => {
                 user: req.user._id
             });
 
+
+
             post.comment.push(comment);
             post.save();
+
+            await comment.populate('post');
+            await comment.populate('user');
+
+            if (req.xhr) {
+
+                return res.status(200).json({
+
+                    data: {
+
+                        comment: comment,
+                        message: "Comment created!"
+                    }
+                });
+            }
+
+            req.flash('success', 'Commented successfuly');
             res.redirect('back');
 
         }
         else {
 
-            console.log('Corresponding post not found in DB');
-
+            req.flash('error', 'Corresponding post not found in DB');
             return res.redirect('back');
         }
 
     } catch (err) {
 
-        console.log('Error: ', err);
+        req.flash('error', err);
 
         return;
     }
@@ -42,26 +60,45 @@ module.exports.destroy = async function (req, res) {
 
         let comment = await Comment.findById(req.params.id).populate('post');
 
-        if (comment.user == req.user.id || comment.post.user == req.user.id) {
+        if (comment) {
 
-            let postId = comment.post;
+            if (comment.user == req.user.id || comment.post.user == req.user.id) {
 
-            await Comment.deleteOne(comment);
-            
-            console.log('Comment deleted.');
+                let postId = comment.post;
 
-            await Post.findByIdAndUpdate(postId, { $pull: { comment: req.params.id } });
-            
+                await Comment.deleteOne(comment);
+
+                req.flash('success', 'Comment deleted.');
+
+                await Post.findByIdAndUpdate(postId, { $pull: { comment: req.params.id } });
+
+                if (req.xhr) {
+
+                    return res.status(200).json({
+
+                        data: {
+
+                            comment_id: req.params.id,
+                            message: "Comment deleted"
+                        }
+                    });
+                }
+
+                return res.redirect('back');
+            }
+            else {
+
+                req.flash('error', 'User not authorized to delete this comment');
+                return res.redirect('back');
+            }
+        } else {
+
+            req.flash('error', 'Corresponding comment not found in DB');
             return res.redirect('back');
         }
-        else {
-
-            console.log('User not authorized to delete this comment');
-
-            return res.redirect('back');
-        }
-
     } catch (err) {
+
+        req.flash('error', err);
 
         console.log('Error: ', err);
