+++ controllers/posts_controller.js
@@ -5,14 +5,30 @@ const Comment = require('../models/comment');
 module.exports.create = async function (req, res) {
 
     try {
-        await Post.create({
+        
+        let post = await Post.create({
             content: req.body.content,
             user: req.user._id
         });
+        
+        await post.populate('user');
+        
+        if(req.xhr){
+
+            return res.status(200).json({
+
+                data: {
+                    post: post,
+                    message: "Post created!"
+                }
+            });
+        }
+
+        req.flash('success', 'Post created successfuly');
 
         return res.redirect('back');
     } catch (err) {
-        console.log('Error: ', err);
+        req.flash('error', err);
     }
 };
 
@@ -26,7 +42,17 @@ module.exports.destroy = async function (req, res) {
 
             await Post.deleteOne(post._id);
 
-            console.log('Post deleted.');
+            if(req.xhr){
+
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: "Post deleted!"
+                })
+            }
+
+            req.flash('success', 'Post deleted.');
 
             await Comment.deleteMany({ post: req.params.id });
 
@@ -36,13 +62,13 @@ module.exports.destroy = async function (req, res) {
         }
         else {
 
-            console.log('User not authorized to delete this post');
+            req.flash('error', 'User not authorized to delete this post');
 
             return res.redirect('back');
         }
     } catch (err) {
 
-        console.log('Error: ', err);
+        req.flash('error', err);
 
         return;
     }
